Moved some GPIO stuff to `stpy`.
